requirements.txt
# ============================
# ✅ Core ML & Data Processing
# ============================
xgboost==1.7.6
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.4
joblib==1.3.2

# ============================
# ✅ Technical Analysis
# ============================
ta==0.10.2

# ============================
# ✅ Web & API
# ============================
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0

# ============================
# ✅ Supabase Client
# ============================
supabase==1.0.3
